requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.5
django-cors-headers==3.4.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-formtools==2.2
django-import-export==2.3.0
django-ranged-response==0.2.0
django-reversion==3.0.7
django-simple-captcha==0.5.12
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.9.2
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==2.0.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.4
Pillow==7.2.0
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.10
xadmin==2.0.1
xlrd==1.2.0
xlwt==1.3.0
